:recycle: refactor: chore: simplify swiper
